Move log in functionality to MainActivity for a cleaner UI flow
