--- app/src/main/java/com/muk/sami/LoginActivity.java
@@ -1,84 +1,0 @@
-package com.muk.sami;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.util.Log;
-
-import com.firebase.ui.auth.AuthUI;
-import com.firebase.ui.auth.ErrorCodes;
-import com.firebase.ui.auth.IdpResponse;
-import com.google.firebase.auth.FirebaseAuth;
-
-import java.util.Arrays;
-
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-
-public class LoginActivity extends AppCompatActivity {
-
-    private static final String TAG = "LoginActivity";
-    private static final int RC_SIGN_IN = 0;
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        signIn();
-
-    }
-
-    private void signIn() {
-        FirebaseAuth auth = FirebaseAuth.getInstance();
-        if (auth.getCurrentUser() != null) {
-            // already signed in
-            startActivity(new Intent(this, MainActivity.class));
-            finish();
-        } else {
-            // not signed in
-            startActivityForResult(
-                    AuthUI.getInstance()
-                            .createSignInIntentBuilder()
-                            .setAvailableProviders(Arrays.asList(
-                                    new AuthUI.IdpConfig.GoogleBuilder().build(),
-                                    new AuthUI.IdpConfig.EmailBuilder().build()))
-                            .setTosAndPrivacyPolicyUrls(
-                                    "https://superapp.example.com/terms-of-service.html",
-                                    "https://superapp.example.com/privacy-policy.html")
-                            .setIsSmartLockEnabled(false)
-                            .build(),
-                    RC_SIGN_IN);
-            overridePendingTransition(0, 0);
-        }
-    }
-
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        if (requestCode == RC_SIGN_IN) {
-            IdpResponse response = IdpResponse.fromResultIntent(data);
-
-            // Successfully signed in
-            if (resultCode == RESULT_OK) {
-                startActivity(new Intent(this, MainActivity.class));
-                finish();
-                overridePendingTransition(0, 0);
-            } else {
-                // Sign in failed
-                if (response == null) {
-                    // User pressed back button
-                    //showSnackbar(R.string.sign_in_cancelled);
-                    finish();
-                    return;
-                }
-
-                signIn();
-
-                if (response.getError().getErrorCode() == ErrorCodes.NO_NETWORK) {
-                    //showSnackbar(R.string.no_internet_connection);
-                    return;
-                }
-
-                //showSnackbar(R.string.unknown_error);
-                Log.e(TAG, "Sign-in error: ", response.getError());
-            }
-        }
-    }
-}
